Use larger integer types in SQL schema (fixes #232)
Some games were missing player info: 4 player games would have
only two players recorded, for instance. What happened is that
the bot games_played field had reached the limit of a MySQL
SMALLINT, so when we tried to increment that field, the database
driver gave us an error, causing the rest of the records to not
be recorded (as the error led to a Python exception).