Use nonblocking pipe to send map to bot
This prevents the game environment from blocking, which is a
potential DoS vector in the competition environment. If a bot does
not read STDIN, but continues sending commands, the game
environment will fill the write pipe between it and the bot,
causing it to block. This isn't counted in the timeout, so the bot
will never be ejected.

This patch changes the write end of the pipe to be nonblocking,
and checks if writing to the pipe would have blocked. If so, it
ejects the bot with a descriptive error message.