Set more reasonable challenge game highliting.
Currently in the game feed the background for challenge games is much too
dark, making it hard to read player names.

On other hand the player profile page sets the challenge game background
to the same color as other games.

This changes to a light cyan color for both pages.