Update distance calculation in get_unsafe_moves
The original distance is calculated by subtracting two points.  Looking at positional.py, we know that by default, this coordinate would be normalized.

So lets say we want to move from `(14,24)` to `(12,24)`.  The distance would be `(46,0)`, which caused the get_unsafe_moves function to think they can go faster by choosing the inverted direction, so we get a direction of `[ (1,0) ]`, but intuitively, we would want `[ (-1,0) ]`

By changing how the distance is calculated, we avoid this situation.